airaad/backend/requirements/base.txt
# =============================================================================
# AirAd Backend — Base Requirements
# Shared across all environments.
# =============================================================================

# Django core
Django==5.1.5
djangorestframework==3.15.2
django-environ==0.11.2

# Auth
djangorestframework-simplejwt==5.3.1

# API documentation
drf-spectacular==0.27.2

# GeoDjango (PostGIS)
# GDAL/GEOS/PROJ installed at OS level in Dockerfile — not via pip

# Async / Task queue
celery==5.4.0
django-celery-beat==2.7.0
redis==5.2.1
kombu==5.3.7

# File storage
django-storages==1.14.4
boto3==1.35.86

# Validation
pydantic==2.10.4

# Cryptography (AES-256-GCM)
cryptography==44.0.0

# CORS
django-cors-headers==4.6.0

# Utilities
Pillow==11.1.0
python-slugify==8.0.4

# Stripe payments
stripe==8.4.0

# Twilio SMS
twilio>=9.0.0

# HTTP client for Google Places API
httpx==0.27.*
